resolve package issues linked to profile settings
