+++ app/src/androidTest/java/com/github/blecoeur/bootcamp/accountSettings/AccountSettingsActivityLoggedInTest.kt
@@ -7,7 +7,6 @@ import android.content.Context
 import android.content.Intent
 import android.net.Uri
 import android.provider.MediaStore
-import androidx.test.core.app.ActivityScenario
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions
@@ -20,8 +19,8 @@ import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.rule.GrantPermissionRule
-import com.github.blecoeur.bootcamp.AccountSettingsActivity
 import com.github.blecoeur.bootcamp.R
+import com.github.blecoeur.bootcamp.profile.settings.AccountSettingsActivity
 import org.junit.*
 import org.junit.runner.RunWith
