+++ app/src/androidTest/java/com/github/blecoeur/bootcamp/accountSettings/AccountSettingsActivityNoPermsTest.kt
@@ -7,8 +7,8 @@ import androidx.test.espresso.Espresso
 import androidx.test.espresso.action.ViewActions
 import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.github.blecoeur.bootcamp.AccountSettingsActivity
 import com.github.blecoeur.bootcamp.R
+import com.github.blecoeur.bootcamp.profile.settings.AccountSettingsActivity
 import org.junit.Test
 import org.junit.runner.RunWith
