+++ app/src/androidTest/java/com/github/blecoeur/bootcamp/accountSettings/AccountSettingsActivityTest.kt
@@ -16,13 +16,10 @@ import androidx.test.espresso.intent.Intents
 import androidx.test.espresso.intent.Intents.*
 import androidx.test.espresso.intent.matcher.IntentMatchers.hasAction
 import androidx.test.espresso.matcher.ViewMatchers.*
-import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.rule.GrantPermissionRule
-import com.github.blecoeur.bootcamp.AccountSettingsActivity
 import com.github.blecoeur.bootcamp.R
-import com.google.firebase.auth.FirebaseAuth
-import org.junit.BeforeClass
+import com.github.blecoeur.bootcamp.profile.settings.AccountSettingsActivity
 import org.junit.Rule
 
 import org.junit.Test
