+++ app/src/main/res/layout/activity_account_settings.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AccountSettingsActivity">
+    tools:context=".profile.settings.AccountSettingsActivity">
 
     <ImageView
         android:id="@+id/profilePic"
